backend/requirements.txt
# Bobby's Workshop - Python Backend Dependencies
# Audio Forensic Intelligence & Stealth Operations

# FastAPI & Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
websockets==14.1

# Audio Processing
librosa==0.10.2
soundfile==0.12.1
scipy==1.14.1
numpy==1.26.4
noisereduce==3.0.0

# AI Transcription
faster-whisper==1.0.3
openai-whisper==20231117  # Fallback if faster-whisper unavailable

# Speaker Diarization
pyannote.audio==3.1.1
pyannote.core==5.0.1

# Voice Activity Detection
webrtcvad==2.0.10

# URL Audio Extraction
yt-dlp==2024.8.6

# Media Processing
ffmpeg-python==0.2.0
Pillow==10.4.0

# USB Hardware Detection (Pandora Codex)
pyusb==1.2.1

# Utilities
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
aiofiles==24.1.0

# Optional: Advanced Neural Processing (Tier 2)
# deepfilternet==0.6.0  # Uncomment when ready for neural dereverberation

# Testing
pytest>=8.3.4
pytest-asyncio>=0.24.0
